+++ packages/babel-plugin-transform-es2015-modules-umd/test/fixtures/umd/export-default-non-function/expected.js
