+++ packages/babel-plugin-transform-es2015-modules-umd/test/fixtures/umd/export-from-as-default/expected.js
@@ -1,0 +1,25 @@
+(function (global, factory) {
+  if (typeof define === "function" && define.amd) {
+    define(["exports", "foo"], factory);
+  } else if (typeof exports !== "undefined") {
+    factory(exports, require("foo"));
+  } else {
+    var mod = {
+      exports: {}
+    };
+    factory(mod.exports, global.foo);
+    global.actual = mod.exports;
+  }
+})(this, function (exports, _foo) {
+  "use strict";
+
+  Object.defineProperty(exports, "__esModule", {
+    value: true
+  });
+  Object.defineProperty(exports, "default", {
+    enumerable: true,
+    get: function () {
+      return _foo.foo;
+    }
+  });
+});
