+++ packages/babel-plugin-transform-es2015-modules-umd/test/fixtures/umd/export-from/actual.js
@@ -1,0 +1,1 @@
+export {foo} from "foo";
