+++ packages/babel-plugin-transform-es2015-modules-umd/test/fixtures/umd/export-from/expected.js
@@ -1,26 +1,25 @@
 (function (global, factory) {
   if (typeof define === "function" && define.amd) {
-    define(["exports"], factory);
+    define(["exports", "foo"], factory);
   } else if (typeof exports !== "undefined") {
-    factory(exports);
+    factory(exports, require("foo"));
   } else {
     var mod = {
       exports: {}
     };
-    factory(mod.exports);
+    factory(mod.exports, global.foo);
     global.actual = mod.exports;
   }
-})(this, function (exports) {
+})(this, function (exports, _foo) {
   "use strict";
 
   Object.defineProperty(exports, "__esModule", {
     value: true
   });
-  exports.foo = foo;
-  exports.foo = foo;
-  exports.bar = bar;
-  exports.bar = foo;
-  exports.default = foo;
-  exports.default = foo;
-  exports.bar = bar;
+  Object.defineProperty(exports, "foo", {
+    enumerable: true,
+    get: function () {
+      return _foo.foo;
+    }
+  });
 });
