+++ packages/babel-plugin-transform-es2015-modules-umd/test/fixtures/umd/export-named-as-default-and-name/actual.js
@@ -1,0 +1,1 @@
+export {foo as default, bar};
