+++ packages/babel-plugin-transform-es2015-modules-umd/test/fixtures/umd/export-named-names/actual.js
@@ -1,0 +1,1 @@
+export {foo, bar};
