+++ packages/babel-plugin-transform-es2015-modules-umd/test/fixtures/umd/export-named/expected.js
@@ -16,16 +16,5 @@
   Object.defineProperty(exports, "__esModule", {
     value: true
   });
-  exports.foo8 = foo8;
-  var foo = exports.foo = 1;
-  var foo2 = exports.foo2 = 1,
-      bar = exports.bar = 2;
-  var foo3 = exports.foo3 = function () {};
-  var foo4 = exports.foo4 = undefined;
-  let foo5 = exports.foo5 = 2;
-  let foo6 = exports.foo6 = undefined;
-  const foo7 = exports.foo7 = 3;
-  function foo8() {}
-  class foo9 {}
-  exports.foo9 = foo9;
+  exports.foo = foo;
 });
