--- packages/babel-plugin-transform-es2015-modules-umd/test/fixtures/umd/exports-from/expected.js
@@ -1,70 +1,0 @@
-(function (global, factory) {
-  if (typeof define === "function" && define.amd) {
-    define(["exports", "foo"], factory);
-  } else if (typeof exports !== "undefined") {
-    factory(exports, require("foo"));
-  } else {
-    var mod = {
-      exports: {}
-    };
-    factory(mod.exports, global.foo);
-    global.actual = mod.exports;
-  }
-})(this, function (exports, _foo) {
-  "use strict";
-
-  Object.defineProperty(exports, "__esModule", {
-    value: true
-  });
-  Object.keys(_foo).forEach(function (key) {
-    if (key === "default" || key === "__esModule") return;
-    Object.defineProperty(exports, key, {
-      enumerable: true,
-      get: function () {
-        return _foo[key];
-      }
-    });
-  });
-  Object.defineProperty(exports, "foo", {
-    enumerable: true,
-    get: function () {
-      return _foo.foo;
-    }
-  });
-  Object.defineProperty(exports, "foo", {
-    enumerable: true,
-    get: function () {
-      return _foo.foo;
-    }
-  });
-  Object.defineProperty(exports, "bar", {
-    enumerable: true,
-    get: function () {
-      return _foo.bar;
-    }
-  });
-  Object.defineProperty(exports, "bar", {
-    enumerable: true,
-    get: function () {
-      return _foo.foo;
-    }
-  });
-  Object.defineProperty(exports, "default", {
-    enumerable: true,
-    get: function () {
-      return _foo.foo;
-    }
-  });
-  Object.defineProperty(exports, "default", {
-    enumerable: true,
-    get: function () {
-      return _foo.foo;
-    }
-  });
-  Object.defineProperty(exports, "bar", {
-    enumerable: true,
-    get: function () {
-      return _foo.bar;
-    }
-  });
-});
