--- packages/babel-plugin-transform-es2015-modules-umd/test/fixtures/umd/exports-named/actual.js
@@ -1,5 +1,0 @@
-export {foo};
-export {foo, bar};
-export {foo as bar};
-export {foo as default};
-export {foo as default, bar};
